Extract getUserInitials utility function
