+++ src/pages/ChatPage.tsx
@@ -4,6 +4,7 @@ import {useAuth} from "@/hooks/useAuth.ts";
 import {Conversation, Message} from "@/types/chat.ts";
 import {BackendDataset} from "@/types/commons.ts";
 import {sendChatMessage} from "@/lib/api.ts";
+import {getUserInitials} from "@/lib/user.ts";
 import dataCommonsIconBlue from '@/assets/data-commons-icon-blue.svg';
 import {Plus, MessageSquare, User, Loader2, Send, ChevronDown, ChevronUp} from "lucide-react";
 import {SearchResultItem} from "@/components/SearchResultItem.tsx";
@@ -44,24 +45,6 @@ const ChatPage: FC = () => {
 
     // To prevent processing initial state multiple times
     const initialQueryProcessed = useRef(false);
-
-    const getInitials = () => {
-        if (!user) return "U";
-        const name = user.name?.trim();
-        if (name) {
-            const names = name.split(/\s+/);
-            if (names.length >= 2) {
-                return (names[0][0] + names[1][0]).toUpperCase();
-            }
-            return name.substring(0, 2).toUpperCase();
-        }
-
-        const preferredUsername = user.preferred_username?.trim();
-        if (preferredUsername) {
-            return preferredUsername.substring(0, 2).toUpperCase();
-        }
-        return user.email ? user.email.substring(0, 2).toUpperCase() : "U";
-    };
 
     const scrollToBottom = () => {
         messagesEndRef.current?.scrollIntoView({behavior: "smooth"});
@@ -602,7 +585,7 @@ const ChatPage: FC = () => {
                                             <div
                                                 className={`w-8 h-8 rounded-full shrink-0 flex items-center justify-center shadow-sm mt-1 overflow-hidden ${msg.sender === 'user' ? 'bg-[#002337] text-white text-sm font-medium' : 'bg-white border border-gray-100 p-1'}`}>
                                                 {msg.sender === 'user' ? (
-                                                    getInitials()
+                                                    getUserInitials(user)
                                                 ) : (
                                                     <img src={dataCommonsIconBlue} alt="Bot"
                                                          className="w-full h-full object-contain"/>
